Fixes a bug that I don't really understand
Trying to complete any 'f' chain resulted in an abort because the 't' chains
are tried first, and when they can't advance, they report REFUSE.

This commit locks the context once a key chain was started.  After that, only
keychains in the locked context are fed new keys until a chain is either
completed or aborted.