Fix non-underlined header bug partially
When the torrent name column (or any other column with a weighted width and no
right-hand side text) is placed on the right side, it's header is not
underlined to the very right. This fixes this issue mostly, but a single space
is still not underlined.

Setting the right-hand side text to any non-space text results in the correct
underlining, but displays the text as expected, so that's not a good
workaround but may help in pinning this bug down.