Add request lock
When the daemon is spammed with multiple requests (which can happen if you open
many tabs with a single command while the connection is already established), it
may not answer some requests and it will eventually timeout.

Not allowing concurrent requests seems to fix this issue, and it doesn't seem to
slow things down too much.