Don't cache the 'files' RPC field
Requesting the 'files' field only when it's necessary (e.g. when initially
requesting files for a torrent or when it or some files in it were renamed) is
too fiddly.  More importantly, if another client renames the torrent or any
files, we don't know about it and keep displaying wrong information.