Fix "RuntimeError: Lock is not acquired."
This was triggered when running an rc file that set one of the unit.* or
unitprefix.* settings and opened lots of tabs.  Not sure exactly how this was
possible, but it seems to have fixed the crash.

Polling only the relevant stuff is also more efficient than simply polling all
the pollers.