Go back to replacing Column/Pile content items in Group.show()
This solves a bug where `tui toggle main` inserts dummy widgets that aren't
removed, decreasing the size of the main area.

Inserting also doesn't make any sense because hide() and show() don't change the
number of items in the Column/Pile. hide() replaces a widget with an empty dummy
widget and show() replaces the dummy with the actual widget.