Do not automatically remove keys from urwid.command_map in bind()
This breaks <ctrl-f>, which is in urwid.command_map to move the cursor to the
right (as in `emacs`), but it is also mapped in the default keymap context to
move down one page in lists and help texts (as in `less`).

Binding keys in urwid.command_map is still possible, but the user has to
explicitly unbind() them first.