Make SettingsAPI items independent
It should now be possible to set and get (from cache or remote service) values
of SettingsAPI items without direct access to the SettingsAPI instance.

This is done by internally forwarding the necessary SettingsAPI methods to
the *RemoteValue instance, which can then fetch new raw values or get them from
a cache and upgrade/parse them.