Normalize unicode in TUI and human-readable CLI only
In principle, the info from the daemon should never be modified. A file name
consists of bytes and may not exist after it went through normalization. This is
an issue when piping output to other commands.

However, composed graphemes (a + ´ = á) mess up rendering because of the
different string length. For now, this is solved with
`unicodedata.normalize('NFC', '...')`. A better approach would be to use
`grapheme` (https://github.com/alvinlindstam/grapheme), but I don't want to add
another dependency for something most users don't need.