Allow specifying torrents by ID with info command in TUI
Torrents can now be selected preferably by focus or preferably by marks. Usually
it's better to prefer marks (e.g. when deleting a bunch of marked torrents), but
in case of the 'summary' command, using the focused torrent is more intuitive.

This fixes a bug where `tab summary id=123` would always show the summary of the
focused torrent.