Fix crash when searching for 'comment' again (#95)
If the torrent list widget is invalidated when setting secondary_filter, it is
redrawn immediately. The list is filtered with the new secondary_filter, but the
new torrent list with the keys needed by secondary_filter hasn't arrived yet. We
apply a filter to Torrents with insufficient keys and key a KeyError.

Registering the new request also polls. We get the update as soon as it's ready,
and then we can apply the new filter as expected.