Refactor SearchableText class
The jump_*() methods now just set a sentinel attribute that is used by render()
to actually do the jump. That's necessary because we need the 'size' variable
for some calculations.

This allowed for the new methods maybe_jump_to_next/prev_match() which only make
the jump if there are no matches on the current page.