Rewrite Timedelta
Timedelta now properly handles years and months when converting to a Timestamp.
For example, the number of seconds in `Timedelta('1y')` depends on the current
time (current year may be leap year), so we ignore the absolute number and
simply add 1 to the number of years with the datetime module.