Default to return value from completion_candidates_posargs()
Do NOT check if its return value evaluates to True because if we called reduce()
on the Candidates object earlier so that nothing matches, it will evaluate to
False, the completer doesn't get it and can't call reduce() to make something
match again.

This is probably only a problem if Candidates objects are cached (e.g. setting
names), but that should be possible for the obvious performance increase.