ui: Print disk space left in the download directory
This commit adds a widget in the bottom bar that displays the amount
of space left in the download directory.

A new poller is implemented that queries the path to the download
directory from the Transmission daemon, and sends it back in a standard
"Free Space" query[1].

If the "incomplete download directory" feature is not disabled and
is different than the regular download directory, it will also be
displayed (separated from the other number with a slash).

Both numbers are pretty printed (e.g. "1KB" instead of "1024"),
and tagged as "important" when below 104 bytes.

[1] https://trac.transmissionbt.com/browser/trunk/extras/rpc-spec.txt#L603

Fixes #146