fix: prevent an infinite loop in `interactive` command
the `interactive` command creates a column layout of `CLIEditWidget`s
and `urwid.Text` widgets. when the user presses tab (shift-tab) focus
should move to the next (previous) `CLIEditWidget`. the loops
responsible for this check for instances of `urwid.Edit`. in b424a1f
`CLIEditWidget` was changed to inherit from `urwid.Wrap` so that it
would have integrated tab completion. this makes the loops in
`interactive` infinite, because no widgets inheriting from `urwid.Edit`
are ever packed into the columns. the fix is to simply update the loop
condition as appropriate. in addition, the helper functions for moving
were simplified.